replace mocha to jest (#17)
* replace mocha with jest
* remove unused function
* edit .npmignore
