+++ jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  bail: true,
+  collectCoverage: true,
+  coverageThreshold: {
+    global: {
+      statements: 100,
+    },
+  },
+};
