+++ lib/Grammar.js
@@ -19,7 +19,6 @@ module.exports = class Grammar {
 
     this.nonterminals = this._getNonterminals();
     this.nonterminalsHash = indexBy(this.nonterminals);
-    this.terminals = this._getTerminals();
 
     this.firstSets = this._makeFirstSets();
     this.followSets = this._makeFollowSets();
@@ -40,20 +39,6 @@ module.exports = class Grammar {
 
   _getNonterminals() {
     return uniq(this.rules.map(({ left }) => left));
-  }
-
-  _getTerminals() {
-    const terminals = [];
-
-    this.rules.forEach((rule) => {
-      rule.right.forEach((item) => {
-        if (this._isTerminal()) {
-          terminals.push(item);
-        }
-      });
-    });
-
-    return terminals;
   }
 
   _isNonterminal(item) {
