+++ package.json
@@ -4,13 +4,13 @@
   "description": "Tool for calculating first, follow and predict sets for a grammar",
   "main": "index.js",
   "scripts": {
-    "test": "mocha --bail ./test/**/*",
+    "test": "jest",
     "lint": "eslint ./",
     "size": "size-limit"
   },
   "size-limit": [
     {
-      "limit": "795 B",
+      "limit": "765 B",
       "path": "index.js"
     }
   ],
@@ -37,10 +37,9 @@
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.19.1",
-    "expect.js": "^0.3.1",
     "husky": "^3.1.0",
+    "jest": "^24.9.0",
     "lint-staged": "^9.5.0",
-    "mocha": "^6.2.2",
     "nrun": "^1.0.0"
   }
 }
