+++ test/.eslintrc.json
@@ -1,5 +1,5 @@
 {
   "env": {
-    "mocha": true
+    "jest": true
   }
 }
